// shared/types/server/fileJson.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Omnition Api Documentation
 * Документация Api для Omnition
 * OpenAPI spec version: 1.0
 */

export interface FileJson {
    /** Название файла */
    name: string;
    /** Путь к файлу */
    filePath: string;
}